(PUP-1151) Factor out v1 HTTP request processing into own request processor.
The end goal is to have v1 and v2 API endpoints in their own processors,
with the routes registered on a shared handler that performs client
authentication and dispatch.

This commit rips out the v1 endpoints to their own processor, but does
not define anything about the handler including registration.

v1-related behavorial tests have been moved from the handler spec to the
v1 spec.