(PUP-1151) Implement Route changes.
Rather than creating a route with a path and method, routes are created
with just a patch matcher and method handlers can be added by calling
the HTTP method-named methods on the route. By default, handlers for all
methods just raise a method not allowed error; user-defined handlers
replace those.

When the handler is given a list of routes to register, it checks to see
if any of them have identical path matching regexes and throws an error
if they do. However, it is still entirely possible for regexes that do
not compare as equal to express the same regular language, so users
still bear responsibility to ensure their routes cooperate.