(PUP-7826) Add metadata loading to tasks.
Previously, tasks could tell you where their metadata file was located,
but not what the file contained. Allow tasks to return their metadata by
adding `read_metadata` and `metadata` methods to the
Puppet::Module::Task. The `read_metadata` method performs the reading
and parsing of the JSON content from the metadata file on disk and
handles the cases where the file is missing or malformed. The `metadata`
method caches the result of `read_metadata` in an instance variable so
the metadata file is only read once for the lifetime of a Task object.